Rename the `contrib.base` submodule to `contrib.studentt`
This is the first step towards adding the equivalent base models with `NormalNoise`.
