+++ calibr8/__init__.py
@@ -6,7 +6,7 @@ from .contrib.noise import (
     StudentTNoise,
     PoissonNoise,
 )
-from .contrib.base import (
+from .contrib.studentt import (
     BaseAsymmetricLogisticT,
     BaseLogIndependentAsymmetricLogisticT,
     BaseModelT,
